Move modint in core
